Introduce asyncio.wait_for to websocket recv
Believe it or not, it was possible for this script to wait to receive
something from the validator for eternity and this is not a good
prroperty in a cron job. `wait_for` will exit after 10 seconds so
that we can move onto the next message. It is hoped 10 is generous
enough, and sufficient enough to generate enough logging to identify
any issues we might otherwise be seeing between validator and
collector nodes.